Added go routes and CORS
